homogeneous_order now works for symbols that are functions
I have put stringent requirements on the functions, perhaps more so than
is really necessary.  The functions must appear in the expression with
exactly the same arguments as provided in the symbols to be checked.
Also, every symbol in the argument of the function must also be a symbol
called in symbols (I am not sure if this is truly necessary, but I want
to play it safe, and so far, only dsolve() uses this, which does nothing
even close to this).