Refactored expand, fixing issues 1455, 1453, and 1445.
Expand hints are now much more specific.  pwoer_exp, power_base, multinomial, mul,
and log have replaced basic as the default hints.  basic remains as a default hint so
it can be used for things that don't fit into those other hints but needs to be automatic.

All expand functions include a recursive keyword argument which determines if it passes
on the expand hints to the inner parts of an expression.

I added expand_mul, expand_log, expand_func, expand_trig, and expand_complex functions
to make it easier to only expand with those hints.

This also changes a few other things internally.  expand_hint functions no longer return
None if they do not expand.  The return the expression.  Mul and Add now have base expand_hint
functions that expands through the terms of the expression.  This replaces similar code
in Basic because not all Basic subclasses works with class.new(*class.args) (e.g., Poly).

This fixes all the tests that it made fail, either by changing the test to an equivalent
expression or by changing the expand calling routine in some code somewhere.

The integration engine relies on expand(log(x**2)) => 2*log(x) working always, even though
it is technically only true when x is positive.  This keeps the old behavior.  Hopefully,
it will be easier to fix when we have the new assumptions system.

Some of this is based on things that I have commited earlier but have not made it in yet.

In particular, this code is based on my patches in Issue 252, which keeps exp(x)*exp(y)
from automatically combining to exp(x+y).  expand_power_exp won't work without it.
It also has some stuff in logcombine and maybe a few other places.