Made constants increment numbers when they are changed
This technically works, but because of the iterative nature of Mul and Add,
the numbers can be incremented too far if the Constant absorbs more than
one term.

This will all probably have to wait until some handler logic is added
to Add and Mul, which would move all of the code into the Constant class,
and would (hopefully) allow constants' numbers to be incremented only once
per absorption.

Until then, I am going to work on a separate branch that uses a function
to absorb constants.  See branch constant-function.