Get all existing tests to pass!
This means the renumbering works (it had to be refactored into a separate function
that is called separately).  Also, one obscure test had to be xfailed (f(C1, C1, x))
because it would be too difficult to make it not return f(C1, C2, x).  dsolve will
never return something like this, and the result is technically correct anyway.