fix 1560 -2**-2 error and standardize names and added a test
This problem did not show up after fixing the problem earlier
because the power calculation code didn't receive an exponent
that was an integer.

Same sort of fix added to handle rational neg base with test.

In both cases, conditions when the exponent is an exact negative integer
are also handled.