Core mul power fixed and tests.
Philosophy: sympy should never return an expression
given in variables that would give a wrong answer when values are substituted
back into that expression.

When there are rational powers of negative bases, those bases cannot be
split up: sqrt(x*y) is not sqrt(x)*sqrt(y) in general unless x and y
are positive, so if the user has not designated them to be thus, they
should not be split apart unless there is some way to add and addition
piece of calculation that will make it true when the values are substituted
back in, e.g. [1, (-1)**-S.Half][(1-sign(x))/2]*sqrt(x)*
              [1, (-1)**-S.Half]((1-sign(y))/2]*sqrt(y)

But that's horrendous (and I haven't verified that it would work).
It seems better to just leave the expression unexpanded and to allow the
signs to take care of themselves. So sqrt(-2*p*n*x) where p is positive,
n is negative and x is unknown in terms of sign would give
sqrt(2)*sqrt(p)*sqrt(-n*x) as the default expansion.

A now superfluous line in _eval_power was removed.

Modifications made to simplify.py's powsimp and separate to incorporate
considerations of the signs of terms being simplified.

A test was added, too.