Added docstrings to all ode.py functions.
This includes adding a docstring to every internal
method, including an explanatory docstring on each hint function
explaining the ODE solving method employed and giving an example.

In addition, I added docstrings to all internal functions so
that you can tell what they do.  I also added an extensive guide
to the module docstring on how to add new ODE solving methods
using the hints engine.

Also created ode.txt in the Sphinx docs that pulls in all of the
relevant docstrings.

This commit also adds simplify key to dsolve(), which lets
you disable constantsimp() and function solving.  Also, dsolve()
now will put the NotImplementedError object in the dictionary with
hint='all' instead of failing if a method raises it.

And I finally removed all of the code from the old module.