Issue 1560b handles neg rational base to neg pwr
This is an extension to issue 1560 except that here, the negative
rational base is being considered. Tests are also added.

There is a small typo that has been fixed (inifinity -> infinity)

The terms base and exp in _eval_power of numbers.py were changed
to b and e to be consistent with other occurences numbers.py.

The term mlib.feq was changed to mlib.mpf_eq; there is no (longer?)
an mlib.feq. A test was added to this which now tests the previously
untested code.