1558: atoms code and docs; NumberSymbol imported; tests
atoms() will take as arguments anything that is loaded in core/__init__.py
By default it returns atoms that are number, symbol or numbersymbol but if
you try to use the argument NumberSymbol alone, it fails because this has
not been imported in core's __init__.py so I made it do so.

There is also an interesting issue of a few tings defined in numbers *not*
being what you think they are: I and zoo do not derive from NumberSymbol, they
derive from Atom (and that can't be changed without breaking things). So
if you want to find them in addition to other number symbols you might try

eq.atoms(I, NumberSymbols)...but this can cause problems since if implicit
things are listed first in atom's args, only those sorts of things are
returned.  I have made changes to make sure that all things given as args
are returned by atoms().

In addition to the above changes, tests have also been added.

/c