real from mpf standardized
When an mpf value is given to Real it should not be stored directly
into number._mpf since it may not be in standard form and then
comparisons between equal numbers will fail because the mpf values
are not the same. This modification uses the passed mpf tuple
to recreate the real which is passed again to real as a real so
the corresponding mpf value will be standardized.