o Real from tuple fix and tests o python_bitcount comments o polys Errors changed to avoid deprecation warning.
o When a Real is pickled, its mpf value is stored with the mantissa in
hex form, but stripped of the 0x part. {It doesn't seem like a good
idea to me to strip that since that number could be confused by
someone later as a long...it does save space, though.} When the object
is loaded again the mpf tuple has a string as the mantissa, e.g.
'13333333333333L'. Previously, this tuple (with string value) was
stored directly in the _mpf record of the object. This is not good
since the mpf tuple usually isn't a string. Also, storing a user
passed value of the tuple is not good because although the tuple might
reconstruct the correct number, if it is not in standard form then
comparisons for equality of that object to another with the same value
but standard mpf tuple will show them as being unequal.

So the behavior of numbers/Real receiving a tuple has been changed so
if the mantissa is a string it will be assumed to be in standard form
and will be converted in hex to a long and that tuple will then be set
as the object's mpf value. If the mantissa is not a string, it will be
assumed that the user has passed this and an mpf value will be
calculated based on the reconstruction of the number from the mpf
tuple, t: ((-1)**t[0] * t[1] * 2**t[2]).

o I added a couple of comments to the python_bitcount routine to make it
clear what is happening.