coeff should be exact match
When asking for the coefficient on some term, coeff sends a
request to collect...but although an expanded expression is
traversed, since the exact flag has not been set to True,
terms with f(x) IN them (rather than being f(x) terms exactly)
are tallied up. This patch passes exact=True to collect. Any
routines depending on the faulty behavior of coeff will need
to be re-written. The portion of ode.py that depended on this
behavior has been rewritten and now all of sympy, as far as
tests are concerned, are compatible with the corrected coeff()
behavior.

small grammar correction make in docstring, too.

The test that identified this problem has been added to the suite.