This concludes the round of ode coverage expansion.
This commit adds test for uncovered cases in
test_undetermined_coefficients_match(), restructures a bit of code with
make_list, and removes an incorrect comment (log(a, b) does NOT have
args[1]).

The total coverage for ode.py is now 99%.  The only parts that are not
tested are code blocks that are there to handle corner cases that might
come up, but for which I do not know of any specific examples that would
cause them to run.  For example, there is a block in dsolve() to
gracefully handle it if a solver raises NotImplementedError with the
'all' hint (it includes the error object in the dictionary instead of
just raising), but of course if a solver ever did this, it would be a
bug and so I would fix that instead of using it to test this feature.
The other parts are similar.

I actually discovered a good number of bugs doing this, so I would
highly recommend that we increase the coverage of all the code.  It's
easiest to do for code that you yourself have written, because you have
to look at an uncovered block and build a test from it.  Hence, I worked
on ode.py.