Fix XFAILing test in test_ode.py
The test is
test_nth_linear_constant_coeff_variation_of_parameters_simplify_False().
Previously, the test was testing the length of the string representation
of a lengthy soltion.  This was a stupid idea of mine.  The important
thing is that the Wronskian should not be simplified in variation of
parameters when simplify=False.  So we just test if the two are
different when simplify=True and when simplify=False.  See commit
4e2056c1, which originally fixed this.

If wronskian() becomes really good at simplifying trig expressions, this
test might fail, but then, the whole thing won't be necessary any more
anyway.