Rewrite heurisch() using Poly arithmetics
This is currently much slower than the Basic arithmetics method, because
of the slowness of DMP multiplication for multinomials with many
generators.  However, the conversion appears to be correct, and should
be faster than the Basic implementation once Poly can multiply
multivariate polynomials together faster (i.e., it uses SDP instead of
DMP).