Start work on solving Risch Differential Equations with weak_normalizer()
Note that Bronstein's book does not have any examples for this
algorithm.  However, based on a few tests, it seems to remove any
positive integer residues from a function.  What remains to be seen is
if it correctly preserves the solution to the RDEs, as predicted in
Theorem 6.1.1 from Bronstein's book.

Also change gcdexdiophantine() to gcdex_diophantine(), and restructure
the code in residue_reduce_derivation() as per
http://asmeurersympy.wordpress.com/2010/06/16/strange-python-behavior-
can-someone-please-explain-to-me-what-is-going-on-here/.