Add NonElementaryIntegral exception
This should provide a more uniform way for subroutines of the Risch
algorithm to indicate to each other when they have proven that the
function being integrated does not have an elementary integral in the
given differential extension.  This should also provide a nice way in
the future to indicate to integrate() that the Risch algorithm has
proven an integral to be non-elementary.

Note that so far this has not been utilized anywhere in the code yet.