No longer give x special status in the Risch Algorithm
The function calls were changed from ..., x, T) to ..., T), where x is
now the first element of T.  This complies with x being the first
element of the any differential extension.  This, along with a typo fix
in special_denom(), fixes a bug that was causing incorrect results from
rischDE().  It also cleans up the code quite a bit, and makes the
recursive nature of the algorithm much clearer.