LambdaPrinter: Don't sort the arguments of Piecewise.
I have previously authored a commit in which I altered all
LambdaPrinter._print_* methods to sort the .args of their arguments.
This has resulted in nonsensical output for Piecewise, in which the
order in which the conditions were specified matters.  This commit
reverts the change to _print_Piecewise, but leaves the sorting intact in
_print_And and _print_Or.