import_module returns a lazy DeferredImport object by default
This object acts just like the module, except it does not actually import the
module until it is used.

The motivation is to prevent external libraries like numpy and matplotlib from
being imported when sympy is imported, which slows down the import time of
sympy.  Note that this still does not achieve this goal because some code in
the plotting module (and possible elsewhere as well) needs to be refactored.

This is still experimental.  This might break under some bizarre
circumstances, though it seems to work for the current code base so far.