Fix box printing
Now instead of complicated (and wrong!) logic, it "merges" together rows using
a key '+' > '|' > '-' > ' ' and just takes the maximum along the two rows
using that key. Horizontal boxes are printed by first transposing and then
printing (a consequence of which is that Box + Box is printed from the top
down).