Refactor the constructor postprocessors
This is a backwards incompatible change. The postprocessors now work like

Mul._constructor_postprocessor_mapping[Class] = [postprocessors]

Only Mul, Add, and Pow are supported. Due to the way class variables work, any
class that wants to support these should add

_constructor_postprocessor_mapping = {}

to its class body.

This should hopefully be faster than the previous way that it was done. If it
isn't, then it likely isn't worth the backwards incompatible change.