Start work on minimal_unsatisfiable_subset()
This is based on code that I originally wrote for conda. See issue #8029.

The code currently doesn't work. We need to refactor satisfiable() and CNF so
that it can take a CNF as input. That way, minimal_unsatisfiable_subset() can
easily bisect the clauses directly, rather than recreating And objects all the
time.