Fix help(latex) to show the docstring
Previously, functools.update_wrapper() was used on the wrapper class instance.
But this added __doc__ to the class instance, which does not work with help()
(help() only looks at __doc__ if it is defined on the class). So we instead
redefine the wrapper class for each decorated function, and use update_wrapper
on the class rather than the instance).

Fixes #20231.